rpm/rpmpm.cc: fix compilation warnings
rpm/rpmpm.cc: In member function 'bool pkgRPMExtPM::ExecRPM(pkgRPMPM::Item::RPMOps, const std::vector<apt_item>&)':
rpm/rpmpm.cc:552:2: warning: this 'if' clause does not guard... [-Wmisleading-indentation]
rpm/rpmpm.cc:554:6: note: ...this statement, but the latter is misleadingly indented as if it were guarded by the 'if'
rpm/rpmpm.cc: In member function 'virtual bool pkgRPMLibPM::Process(const std::vector<apt_item>&, const std::vector<apt_item>&, const std::vector<apt_item>&)':
rpm/rpmpm.cc:993:61: warning: cast to pointer from integer of different size [-Wint-to-pointer-cast]

+++ apt-pkg/rpm/rpmpm.cc
@@ -555,7 +555,7 @@ bool pkgRPMExtPM::ExecRPM(Item::RPMOps op, const std::vector<apt_item> &files)
     {
 	if (!cmd.empty())
 	    cmd += ' ';
-	    cmd += Args[i];
+	cmd += Args[i];
     }
 
    bool FilesInArgs = true;
@@ -844,7 +844,7 @@ bool pkgRPMLibPM::Process(const std::vector<apt_item> &install,
    rpmReadConfigFiles(NULL, NULL);
 
    int probFilter = 0;
-   int notifyFlags = 0;
+   unsigned int notifyFlags = 0;
    int tsFlags = 0;
 
    if (uninstall.empty() == false)
@@ -1006,11 +1006,13 @@ bool pkgRPMLibPM::Process(const std::vector<apt_item> &install,
    probFilter |= rpmtsFilterFlags(TS);
    rpmtsSetFlags(TS, (rpmtransFlags)(rpmtsFlags(TS) | tsFlags));
    rpmtsClean(TS);
-   rc = rpmtsSetNotifyCallback(TS, rpmShowProgress, (void *)notifyFlags);
+   rc = rpmtsSetNotifyCallback(TS, rpmShowProgress,
+                               (void *) (unsigned long) notifyFlags);
    rc = rpmtsRun(TS, NULL, (rpmprobFilterFlags)probFilter);
    probs = rpmtsProblems(TS);
 #else
-   rc = rpmRunTransactions(TS, rpmShowProgress, (void *)notifyFlags, NULL,
+   rc = rpmRunTransactions(TS, rpmShowProgress,
+                           (void *) (unsigned long) notifyFlags, NULL,
                            &probs, (rpmtransFlags)tsFlags,
 			   (rpmprobFilterFlags)probFilter);
 #endif
